Fix the solution to exercise 9.4 to work with nested terms
The previous solution was producing wrong answer to the following query:

```prolog
?- groundterm(foo(bar(_))).
true.
```

It is important to note that `groundterm_in_list` should be called in
`groundterm` with a list that does not consist the functor, otherwise
we would end up in an infinite loop. For example:

```prolog
?- groundterm(1).
Out of stack error
```