Update Exercise 6.5 to not use helper functions
For swapfl assumption is made that empty list and list with one
element will return false given they do not have first and last
element. On this assumption base case is made with two elements
swapped, and inductive case with head and tail being different
but elements in between are same.

This allows to define swapfl in terms of itself, without any extra
helper functions.